regex: match zero or more ANSI escapes, not just 0 or 1
This causes a temporary test failure in which blank lines are left
unmarked. This is due to a now-exposed bug in
"strip_leading_indicators". That logic looked for a color escape
sequence, but did not recognize multiple. The output of mark_empty_line
includes multiple escape sequences; and so before those would be passed
through and now are not.

However, it is intentional that we should allow multiple escapes: there
is nothing weird about "set background; set color; set bold" sent as
multiple sequences, and those should be preserved as though one "color".